# first pass on optimization of nif parsing functions from the file stream