--- a/components/files/constrainedfilestream.cpp
+++ b/components/files/constrainedfilestream.cpp
@@ -8,7 +8,7 @@
 namespace
 {
 // somewhat arbitrary though 64KB buffers didn't seem to improve performance any
-const size_t sBufferSize = 4096;
+const size_t sBufferSize = 8192;
 }
 
 namespace Files