… components/nif/nifstream.cpp