This pull request could be used as is but I also want to start the discussion on further optimization.

I've done a few things in this changeset to help performance:
1. Hoisted all of the functions from nifstream.cpp to nifstream.hpp so that they can be considered for inlining by the compiler.
2. Removed the manual little endian massaging of the stream and instead do direct casting of the buffer read to the target data type.  This will break portability to big endian architectures.  Not sure if that was a target but I can always put it as a fallback if not on an x86 architecture.
3.  Pre-read the size of the entire structure from the buffer to reduce calls to inp->read().
4.  Doubled the size of the file buffer.

Using google benchmark on multiple iterations of parsing base_anim.nif, I get this from the original code:
---------------------------------------------------------
Benchmark                  Time           CPU Iterations
---------------------------------------------------------
^[[m^[[0;32mBM_NifRead_mean   ^[[m^[[0;33m  28403072 ns   28390625 ns ^[[m^[[0;36m        25^[[m
^[[m^[[0;32mBM_NifRead_median ^[[m^[[0;33m  28286460 ns   28125000 ns ^[[m^[[0;36m        25^[[m
^[[m^[[0;32mBM_NifRead_stddev ^[[m^[[0;33m    769761 ns     811667 ns ^[[m^[[0;36m        25^[[m
^[[m

After my optimizations, I get this:
---------------------------------------------------------
Benchmark                  Time           CPU Iterations
---------------------------------------------------------
^[[m^[[0;32mBM_NifRead_mean   ^[[m^[[0;33m  21612806 ns   21594238 ns ^[[m^[[0;36m        32^[[m
^[[m^[[0;32mBM_NifRead_median ^[[m^[[0;33m  21431406 ns   21484375 ns ^[[m^[[0;36m        32^[[m
^[[m^[[0;32mBM_NifRead_stddev ^[[m^[[0;33m    580247 ns     639873 ns ^[[m^[[0;36m        32^[[m

Just taking the mean, we get about a 30% speedup for this.  

Other things I want to consider is not using the traditional istream read and instead provide a funciton to move data directly from the internal 8K buffer into the target data type (uint16_t,uint32_t,uint64_t,osg::Vec3f,osg::Quat).

This will cut the move requirements from:
istream buffer -> intermediate buffer -> target data type
to this:
istream buffer -> target data type

Also thought using some SIMD intrinsics or even a direct memcopy may be quicker than some of these go between functions we use for the types.  These efforts might eek out another 10% or so.

Thoughts?