@psi29a 2017-09-21
This will break on Debian and Ubuntu big endian systems, the packages for those systems are built for without asking. We want to be as portable as we can reasonably be. 

Please add a fallback as you suggest.