@kcat 2017-09-21
> Removed the manual little endian massaging of the stream and instead do direct casting of the buffer read to the target data type. This will break portability to big endian architectures. Not sure if that was a target but I can always put it as a fallback if not on an x86 architecture.

I'd suggest not to break it. I don't know what the status of OpenMW on big-endian is, but it'd be good to not break it if possible. You should be able to make an inline wrapper function which does a simple casting on little-endian systems and leaves the byte swapping to only occur on big-endian using a compile-time check.

> Pre-read the size of the entire structure from the buffer to reduce calls to inp->read().

Certain structures have different sizes depending on values in the structure. There may be flags indicating the presence (or absence) of fields, and some contain a count of items to read.