@scrawl 2017-09-21
Regarding the benchmark, animation files like base_anim are probably a worst-case scenario, with many small fields to read, but this file is only loaded once by the engine. Typical mesh files with vertices/triangles can be read in big chunks with not much effort. Any chance we could get a benchmark for an actual game session?