@crussell187 2017-09-21
> This will break on Debian and Ubuntu big endian systems, the packages for those systems are built for without asking. We want to be as portable as we can reasonably be.

What are the list of architectures we need support?  I'll explore some ways to enable it on ARM if I detect the compilation target for ARM is little endian.  In the current practical use case,  a significant portion of our use base is x86 and will get the benefits.  

> This isn't standard. Do you really need SSE?

This was included to get both compilers to include __m128.  Now with more research, I am finding that __m128 is not supported universally so I will find another portable replacement.  This data type is to hint to the compiler to use that big 128 bit register file in the machine sitting there doing nothing to move data larger than 64 bits (Vector3, Vector4, Matrix, etc.)

> Certain structures have different sizes depending on values in the structure. There may be flags indicating the presence (or absence) of fields, and some contain a count of items to read.

I agree with you but a lot of calls I saw in nif code were of the following:
if(nif->getInt()){
   foo = nif->getFloat();
   foo2 = nif->getInt();
   foo3 = nif->getMatrix3();
}
All of these can be amortized with buffer read.  That being said for readability of the nif code I would prefer to go this route to reduce the overhead of the istream read calls():

> Other things I want to consider is not using the traditional istream read and instead provide a function to move data directly from the internal 8K buffer into the target data type (uint16_t,uint32_t,uint64_t,osg::Vec3f,osg::Quat).