@crussell187 2017-09-21
> Regarding the benchmark, animation files like base_anim are probably a worst-case scenario, with many small fields to read, but this file is only loaded once by the engine. Typical mesh files with vertices/triangles can be read in big chunks with not much effort. Any chance we could get a benchmark for an actual game session?

I'll do some isolated benchmarks first with the other file types.  Although those nifs are streaming more uniform data, I still expect some improvements due to the little endian massaging, lack of inlining, and lack of istream read amortization for Vector3 calls.

Do we have a good way of benchmarking in game?  I would like to perform a worst case asset load.  What part of morrowind is the worst? :)