@psi29a 2017-09-21
To give you an example, here is the openal-soft (from kcat/chris) package that I maintain on Debian (and Ubuntu):
https://buildd.debian.org/status/package.php?p=openal-soft
That's 23 different archs. 

https://buildd.debian.org/status/package.php?p=openmw
Not all are supported because we're not yet in main. That will happen soon.

Please don't just drop BE support because of gaining a few ms.