@kcat 2017-09-21
> This was included to get both compilers to include __m128. Now with more research, I am finding that __m128 is not supported universally so I will find another portable replacement.

__m128 is part of the SSE intrinsics, and not all systems support SSE. Anything beyond 64-bit data types requires system- and compiler-specific extensions, which is likely to be an unnecessary optimization (too much work/maintenance for too little gain). I bet you could find more speed improvements by loading the whole file into memory first, then read the individual elements from there (kind of like how we handle save games, but in reverse).