@scrawl 2017-09-21
> Do we have a good way of benchmarking in game? I would like to perform a worst case asset load. What part of morrowind is the worst? :)

Running with `--skip-menu --start=balmora` might be a good test.

Note that most of the NIF loading happens from background threads and is most likely going to be on-time anyway. Therefore I'm not overly keen on dodgy optimizations that make the code harder to read and maintain. SSE sounds like overkill, and from what I've heard doesn't gain you much for code that isn't data-oriented. We should stick to changes that have a measured benefit. You've measured the overall speedup from this branch, but can you list, for the 4 improvements in your original post, the impact of each change respectively?