@crussell187 2017-09-21
> Note that most of the NIF loading happens from background threads and is most likely going to be on-time anyway. Therefore I'm not overly keen on dodgy optimizations that make the code harder to read and maintain. SSE sounds like overkill, and from what I've heard doesn't gain you much unless you're using it to batch calculations. We should stick to changes that have a measured benefit. You've measured the overall speedup from this branch (using a release build I hope), but can you list, for the 4 improvements in your original post, the impact of each change respectively?

Latest gcc with -O3.  I wouldn't care to test this on debug (inlining wouldn't even be considered there.)

I would like to replace the intent of SSE specifically with just SIMD.  If the inlining is done correctly and call overhead is removed, you can do faster data movement from the file buffer to your data structures with the SIMD moves.  The only thing faster will be a direct mem copy for large data structures.  There is a portable way to do this and I will update my pull to reflect that.

I'll prepare some benchmarks with each separate change.