@crussell187 2017-09-22
So I'll eat my words on my conclusions about SIMD.  I don't think the operations made a difference to the code.  This new commit has all of the SIMD taken out and a portable ifdef to only rely on little endian alignment on an x86 architecture.  Otherwise, it does the big endian conversion from the nif.  Below is the benchmark for the base_anim.nif with the new code:

---------------------------------------------------------
Benchmark                  Time           CPU Iterations
---------------------------------------------------------
BM_NifRead_mean     20961216 ns   20947266 ns         32
BM_NifRead_median   20771219 ns   20996094 ns         32
BM_NifRead_stddev     612575 ns     652292 ns         32

Slightly faster than my initial optimization.  Hopefully this code seems much more portable now.  I'll post some more benchmarks soon.