@scrawl 2017-09-22
Great! The optimization even removes more lines than it added :)

I wouldn't call bloodsplat.nif a typical file though (it's a particle effect, which use lots of fragmented data also).