@MiroslavR 2017-09-22
Can you make the endianness checks reusable (e.g. a macro in endian.hpp)?